+++ src/users/user.routes.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express';
 import prisma from '../common/prisma/client';
-import UserRepository from './user-repository';
-import UserService from './user-service';
-import UserController from './user-controller';
+import UserRepository from './user.repository';
+import UserService from './user.service';
+import UserController from './user.controller';
 import validateDto from '../common/utils/validate.dto';
 import { CreateUserDto } from './dtos/create-user.dto';
